EN:
- added saving of user settings (language, theme and push notifications) in the device memory so that they are applied after starting the application
- added provider to handle push notifications
- added shared preferences to store user settings
- wrapping the language selection on the settings screen in an additional container to make it visually match the rest of the settings a bit better
- improved logout (now we correctly make a request to the API to remove the token from the database)
- minor visual fixes and expansions of theme settings

PL:
- dodanie zapisywania ustawień użytkownika (język, motyw oraz powiadomienia push) w pamięciu urządzenia tak by były zastosowane po uruchomieniu aplikacji
- dodanie providera do obsługi powiadomień push
- dodanie shared preferences do przetrzymywania ustawień użytkownika
- opakowanie wyboru językana ekranie ustawień  w dodatkowy kontener aby nieco lepiej wizualnie pasowało do reszty ustawień
- poprawienie wylogowywania (teraz poprawnie wykonujemy request do api aby usunąć token z bazy danych)
- drobne poprawki wizualne i rozwinięcia ustawień motywów

+++ lib/data/services/auth_service.dart
@@ -15,13 +15,10 @@ class AuthService {
 
   Future<bool> login(String username, String password) async {
     try {
-      final response = await _apiService.dio.post(
-        '/auth/login',
-        data: {
-          'username': username,
-          'password': password,
-        }
-      );
+      final response = await _apiService.dio.post('/auth/login', data: {
+        'username': username,
+        'password': password,
+      });
 
       if (response.statusCode == 201 && response.data['access_token'] != null) {
         final accessToken = response.data['access_token'];
@@ -76,9 +73,22 @@ class AuthService {
     return null;
   }
 
-  Future<void> logout() async {
-    await _secureStorage.delete(key: 'access_token');
-    await AuthStorage.deleteLoginData();
+  Future<bool> logout() async {
+    try {
+      final response = await _apiService.dio.post(
+        '/auth/logout',
+        options: Options(extra: {'withCredentials': true}),
+      );
+      if (response.statusCode == 200) {
+        await _secureStorage.delete(key: 'access_token');
+        await AuthStorage.deleteLoginData();
+        return true;
+      }
+      return false;
+    } catch (e) {
+      debugPrint("Logout failed: $e");
+      return false;
+    }
   }
 
   Future<String?> getAccessToken() async {
